# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HarvestCore API - Commodity-backed tokenization and lending platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, active, repaid, defaulted, in_liquidation]
        - name: borrowerId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Loans list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"
    post:
      operationId: createLoanRequest
      tags: [loans]
      summary: Submit a loan request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
      responses:
        "201":
          description: Loan request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan details
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loan detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/approve:
    post:
      operationId: approveLoan
      tags: [loans]
      summary: Approve loan (Lender)
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveLoanRequest"
      responses:
        "200":
          description: Loan approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/reject:
    post:
      operationId: rejectLoan
      tags: [loans]
      summary: Reject loan
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Loan rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/repay:
    post:
      operationId: repayLoan
      tags: [loans]
      summary: Record a repayment
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentRequest"
      responses:
        "200":
          description: Repayment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/ltv:
    get:
      operationId: getLoanLtv
      tags: [loans]
      summary: Get current LTV for a loan
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LTV details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LtvReport"
  /loans/{loanId}/risk:
    get:
      operationId: getLoanRisk
      tags: [loans]
      summary: Get risk profile for borrower
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskProfile"

  # ── MARKETPLACE ────────────────────────────────────────────
...
components:
  schemas:
...
    Loan:
      type: object
      properties:
        id:
          type: string
        borrowerId:
          type: string
        borrowerName:
          type: string
        lenderId:
          type: string
          nullable: true
        lenderName:
          type: string
          nullable: true
        tokenId:
          type: string
        commodity:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        tenureDays:
          type: integer
        ltv:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, approved, active, repaid, defaulted, in_liquidation]
        riskScore:
          type: string
          enum: [low, medium, high]
          nullable: true
        disbursedAt:
          type: string
          format: date-time
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        repaidAt:
          type: string
          format: date-time
          nullable: true
        outstandingBalance:
          type: number
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, borrowerId, borrowerName, tokenId, commodity, principalAmount, interestRate, tenureDays, status, createdAt]
    LoanListResponse:
      type: object
      properties:
        loans:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [loans, total, page, limit]
    CreateLoanRequest:
      type: object
      properties:
        tokenId:
          type: string
        principalAmount:
          type: number
        tenureDays:
          type: integer
        purpose:
          type: string
      required: [tokenId, principalAmount, tenureDays]
    ApproveLoanRequest:
      type: object
      properties:
        interestRate:
          type: number
        conditions:
          type: string
      required: [interestRate]
    RepaymentRequest:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [mpesa, bank_transfer, stablecoin, pesalink]
        transactionRef:
          type: string
      required: [amount, paymentMethod]
    LtvReport:
      type: object
      properties:
        loanId:
          type: string
        currentLtv:
          type: number
        outstandingBalance:
          type: number
        collateralValue:
          type: number
        ltvStatus:
          type: string
          enum: [healthy, monitoring, margin_call, liquidation]
        commodityPrice:
          type: number
        priceLastUpdated:
          type: string
          format: date-time
      required: [loanId, currentLtv, outstandingBalance, collateralValue, ltvStatus, commodityPrice, priceLastUpdated]
    RiskProfile:
      type: object
      properties:
        userId:
          type: string
        riskScore:
          type: string
          enum: [low, medium, high]
        creditScore:
          type: integer
        dscr:
          type: number
        isOnNegativeList:
          type: boolean
        repaymentHistory:
          type: string
          enum: [excellent, good, fair, poor]
        assessedAt:
          type: string
          format: date-time
      required: [userId, riskScore, creditScore, dscr, isOnNegativeList, repaymentHistory, assessedAt]

    # ── MARKETPLACE ────────────────────────────────────────────
...